backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
PyJWT==2.8.0
PyOTP==2.9.0
qrcode==7.4.2
Pillow==9.5.0
python-decouple==3.8
django-environ==0.11.2
gunicorn==21.2.0
